chore(app): move work manager works to separate module
